+++ mitra_models/src/posts/helpers.rs
@@ -4,6 +4,7 @@ use crate::{
     accounts::types::{Permission, User},
     bookmarks::queries::find_bookmarked_by_user,
     conversations::queries::{
+        find_moderated_conversations_by_user,
         find_tracking_statuses_by_user,
         is_conversation_participant,
     },
@@ -89,6 +90,8 @@ pub async fn add_user_actions(
     let hidden_posts = find_posts_hidden_by_user(db_client, user_id, &posts_ids).await?;
     let tracking_statuses =
         find_tracking_statuses_by_user(db_client, user_id, &posts_ids).await?;
+    let moderated_conversations =
+        find_moderated_conversations_by_user(db_client, user_id, &posts_ids).await?;
     let get_actions = |post: &PostDetailed| -> PostActions {
         let liked = reactions.iter()
             .any(|(post_id, content)| *post_id == post.id && content.is_none());
@@ -106,6 +109,8 @@ pub async fn add_user_actions(
         let maybe_tracking_status = tracking_statuses.iter()
             .find(|(post_id, _)| *post_id == post.id)
             .and_then(|(_, tracking_status)| *tracking_status);
+        let can_moderate_conversation =
+            moderated_conversations.contains(&post.id);
         PostActions {
             liked: liked,
             reacted_with: reacted_with,
@@ -114,6 +119,7 @@ pub async fn add_user_actions(
             voted_for: voted_for,
             hidden: hidden,
             conversation_tracking_status: maybe_tracking_status,
+            can_moderate_conversation: can_moderate_conversation,
         }
     };
     for post in posts {